+++ src/theme/components/card.ts
@@ -43,6 +43,7 @@ const Card = {
     const {
       hasImage,
       imageIsAtEnd,
+      isAlignedRightActions,
       isBordered,
       isCentered,
       layout,
@@ -86,12 +87,14 @@ const Card = {
       flexFlow: "column wrap",
       textAlign: isCentered ? "center" : null,
       actions: {
-        width: ["100%", "100%", "180px"],
-        marginLeft: ["0", "0", "m"],
-        marginTop: ["xs", "xs", "0"],
+        flexShrink: { base: isAlignedRightActions ? "0" : null, md: "0" },
+        marginLeft: { base: "0", md: "m" },
+        marginTop: { base: "xs", md: "0" },
+        width: { base: "100%", md: "180px" },
       },
       body: {
         display: { md: "block" },
+        flexBasis: { sm: isAlignedRightActions ? "100%" : null, md: "auto" },
         flexFlow: { md: "row nowrap" },
         margin: bodyMargin,
         padding: bodyPadding,